# api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Oxmail API
  version: 1.0.0
  description: Oxmail mail server management API
servers:
  - url: http://localhost:8080
    description: Development server
paths:
  /api/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Authenticate and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change user password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/mail/signature/{email}:
    get:
      operationId: getMailSignature
      tags: [mail]
      summary: Get webmail signature settings for a mailbox
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        "200":
          description: Webmail signature settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signature"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      operationId: upsertMailSignature
      tags: [mail]
      summary: Save webmail signature settings for a mailbox
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignatureRequest"
      responses:
        "200":
          description: Webmail signature settings saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signature"
        "400":
          $ref: "#/components/responses/BadRequest"
    delete:
      operationId: deleteMailSignature
      tags: [mail]
      summary: Delete webmail signature settings for a mailbox
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        "204":
          description: Webmail signature settings deleted
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/contacts/{userEmail}:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts for a mailbox
      parameters:
        - name: userEmail
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        "200":
          description: Contacts for mailbox
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact for a mailbox
      parameters:
        - name: userEmail
          in: path
          required: true
          schema:
            type: string
            format: email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactRequest"
      responses:
        "201":
          description: Contact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/contacts/{userEmail}/{contactID}:
    put:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: userEmail
          in: path
          required: true
          schema:
            type: string
            format: email
        - name: contactID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactRequest"
      responses:
        "200":
          description: Contact updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: userEmail
          in: path
          required: true
          schema:
            type: string
            format: email
        - name: contactID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Contact deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    AuthSession:
      type: object
      required: [email, role, mustChangePassword]
      properties:
        email:
          type: string
        role:
          type: string
          enum: [admin, user]
        mustChangePassword:
          type: boolean
    ChangePasswordRequest:
      type: object
      required: [email, currentPassword, newPassword]
      properties:
        email:
          type: string
          format: email
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 8
    StatusResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
...
    Signature:
      type: object
      required: [email, content, enabled]
      properties:
        email:
          type: string
          format: email
        content:
          type: string
        enabled:
          type: boolean
    SignatureRequest:
      type: object
      required: [content, enabled]
      properties:
        content:
          type: string
        enabled:
          type: boolean
    Contact:
      type: object
      required: [id, userEmail, name, email, createdAt]
      properties:
        id:
          type: integer
          format: int64
        userEmail:
          type: string
          format: email
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateContactRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
    UpdateContactRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    LoginResponse:
      type: object
      required: [status, email, mustChangePassword]
      properties:
        status:
          type: string
          enum: [ok]
        email:
          type: string
        role:
          type: string
          enum: [admin, user]
        mustChangePassword:
          type: boolean
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []